Get all variable types in one go
Rather than sending each signature one at a time to QuickSpec to check
for types without any variables, we now batch those statements together
into one big "main" function so it only takes one nix-eval call.

This means that we should always perform either zero nix-eval
invocations (if we have no clusters, or the clusters are empty), or
two (one to get all of the variable types, one to do all of the
exploring).

This makes the cost of using nix-eval constant, regardless of how many
clusters we have.