Bump version, and depend on bugfixed mlspec-helper
Our Haskell package doesn't actually import anything from mlspec-helper, but
running the executable does rely on mlspec-helper being installed into the
nix-eval environment.

If that version is old, it may have lingering bugs, which we won't spot until
runtime. To prevent that, our executable now depends on mlspec-helper with a
version bound that includes bug fixes. That way, the build fails if a new enough
version of mlspec-helper isn't in the environment.