Use 'fail' to abort parse rather than 'error'
If 'error' gets hit, there's no way to get any context. By using 'fail' we get a
`Left err` case to play with.

In particular, we first try to parse a list of clusters, and if that fails we
try to parse a single cluster. By using 'error', we were only shown the reason
why the latter was failing; but that's a fallback case anyway! It was hence
complaining that '[...]' isn't an object, which is a perfectly legitimate
criticism but clearly, since we're passing in arrays, we *intended* for the
first attempt to match. The reason why it failed (and hence we fell back to the
second attempt) was lost.

Now it's not lost: we show both. In particular, we can see an informative
message like 'type should be String, got Null' as the reason the first attempt
failed.